requirements.txt
numpy==1.26.3
streamlit==1.30.0
tensorflow-macos==2.15.0
